feat: introduce CheckoutConfiguration and related types for subscription management
